# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: TaxFlow 360 API
  version: 1.0.0
servers:
  - url: https://api.example.invalid/v1
paths:
...
  /companies/{companyTaxId}/digital-twin/latest:
    get:
      operationId: getLatestDigitalTwinProjection
      security:
        - oauth2: [digital-twin.read]
      parameters:
        - $ref: '#/components/parameters/CompanyTaxId'
        - {$ref: '#/components/parameters/DigitalTwinScenarioId'}
      responses:
        '200':
          description: Latest authorized projection summary
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DigitalTwinProjection'}
        '404': {$ref: '#/components/responses/Problem'}
  /companies/{companyTaxId}/digital-twin/history:
    get:
      operationId: getDigitalTwinHistory
      security:
        - oauth2: [digital-twin.read]
      parameters:
        - $ref: '#/components/parameters/CompanyTaxId'
        - {name: from, in: query, required: false, schema: {type: string, format: date}}
        - {name: to, in: query, required: false, schema: {type: string, format: date}}
      responses:
        '200':
          description: Immutable projection history
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DigitalTwinProjection'}}
        '404': {$ref: '#/components/responses/Problem'}
...
  /companies/{companyTaxId}/shadow-tax:
    get:
      operationId: listShadowTaxDivergences
      security: [{oauth2: [reconciliation.read]}]
      parameters:
        - {name: companyTaxId, in: path, required: true, schema: {type: string, pattern: '^\d{14}$'}}
        - {$ref: '#/components/parameters/TenantId'}
      responses:
        '200':
          description: Tenant-scoped immutable divergence history
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ShadowTaxDivergence'}}}}
        '403': {$ref: '#/components/responses/Problem'}
  /companies/{companyTaxId}/shadow-tax/{divergenceId}:
    get:
      operationId: getShadowTaxDivergence
      security: [{oauth2: [reconciliation.read]}]
      parameters:
        - {name: companyTaxId, in: path, required: true, schema: {type: string}}
        - {name: divergenceId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/TenantId'}
      responses:
        '200': {description: Divergence evidence, content: {application/json: {schema: {$ref: '#/components/schemas/ShadowTaxDivergence'}}}}
        '404': {$ref: '#/components/responses/Problem'}
...
  /regulatory/authorities:
    get:
      operationId: listRegulatoryAuthorities
      security: [{oauth2: [regulatory.read]}]
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Approved official authorities and canonical origins
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/RegulatoryAuthority'}}}}
  /regulatory/answer:
    post:
      operationId: answerRegulatoryQuestion
      security: [{oauth2: [regulatory.read]}]
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question, cutoff_at, authority_ids, document_types, jurisdiction]
              properties:
                question: {type: string, minLength: 1}
                cutoff_at: {type: string, format: date-time}
                authority_ids: {type: array, minItems: 1, items: {type: string}}
                document_types: {type: array, minItems: 1, items: {type: string}}
                jurisdiction: {type: string}
      responses:
        '200': {description: Cited answer, conflict, or explicit refusal, content: {application/json: {schema: {$ref: '#/components/schemas/RegulatoryAnswer'}}}}
        '403': {$ref: '#/components/responses/Problem'}
  /regulatory/changes/{changeId}:
    get:
      operationId: getRegulatoryChangeHistory
      security: [{oauth2: [regulatory.read]}]
      parameters:
        - {name: changeId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/TenantId'}
      responses:
        '200': {description: Immutable change history, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/RegulatoryChange'}}}}}
        '404': {$ref: '#/components/responses/Problem'}
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://identity.example.invalid/oauth2/token
          scopes:
            transactions.write: Submit transactions
            readiness.read: Read readiness results
            simulations.write: Create tax simulations
            simulations.read: Read tax simulations and comparisons
            digital-twin.read: Read Digital Twin projections and evidence
            reconciliation.read: Read tenant-scoped Shadow Tax divergences
            reconciliation.review: Decide authorized reconciliation reviews
            regulatory.read: Search and read cited regulatory evidence
            regulatory.review: Submit and approve regulatory change requests
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    CompanyTaxId:
      name: companyTaxId
      in: path
      required: true
      schema: {type: string, minLength: 11, maxLength: 14}
    DigitalTwinScenarioId:
      name: scenarioId
      in: query
      required: false
      schema: {type: string, default: baseline}
...
  responses:
    Problem:
      description: RFC 9457 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    OfficialSourceReference:
      type: object
      additionalProperties: false
      required: [sourceId, sourceUrl, authority, documentId, provision, publicationDate, capturedAt, contentSha256, status]
      properties:
        sourceId: {type: string, minLength: 1}
        sourceUrl: {type: string, format: uri, pattern: '^https://'}
        authority: {type: string, minLength: 1}
        documentId: {type: string, minLength: 1}
        provision: {type: string, minLength: 1}
        publicationDate: {type: string, format: date}
        capturedAt: {type: string, format: date-time}
        contentSha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        status: {type: string, const: validated}
...
    DigitalTwinEvidence:
      type: object
      additionalProperties: false
      required: [evidenceId, evidenceType, driver, sourceDataset, sourceRecordId, observedValue, contribution]
      properties:
        evidenceId: {type: string}
        evidenceType: {type: string, enum: [financial_fact, stress_assumption, tax_simulation, model_metric]}
        driver: {type: string}
        sourceDataset: {type: string}
        sourceRecordId: {type: string}
        simulationId: {type: string, description: Required for tax evidence}
        ruleId: {type: string, description: Inherited from the immutable tax simulation}
        sourceIds: {type: array, items: {type: string}, description: Official citation IDs inherited from Wave 3}
        observedValue: {type: string}
        contribution: {type: string}
    DigitalTwinModelEvidence:
      type: object
      additionalProperties: false
      required: [mode, trainingWindowStart, trainingWindowEnd, mape, absoluteBias, intervalCoverage]
      properties:
        mode: {type: string, enum: [deterministic_baseline, promoted_forecast]}
        modelVersion: {type: string}
        trainingWindowStart: {type: string, format: date}
        trainingWindowEnd: {type: string, format: date}
        mape: {type: string, pattern: '^\d+\.\d+$'}
        absoluteBias: {type: string, pattern: '^\d+\.\d+$'}
        intervalCoverage: {type: string, pattern: '^\d+\.\d+$'}
        promotionActor: {type: string}
        fallbackReason: {type: string}
...
    DigitalTwinProjection:
      type: object
      additionalProperties: false
      required: [projectionId, companyTaxId, scenarioId, cutoffAt, publishedAt, horizonDays, assumptionVersion, modelEvidence, minimumClosingCash, maximumLiquidityGap, taxFloatDelta, workingCapitalNeed, evidence]
      properties:
        projectionId: {type: string}
        companyTaxId: {type: string}
        scenarioId: {type: string}
        cutoffAt: {type: string, format: date-time}
        publishedAt: {type: string, format: date-time}
        horizonDays: {type: integer, enum: [30, 90, 180, 365]}
        assumptionVersion: {type: string}
        modelEvidence: {$ref: '#/components/schemas/DigitalTwinModelEvidence'}
        minimumClosingCash: {$ref: '#/components/schemas/SignedDecimalMoney'}
        maximumLiquidityGap: {$ref: '#/components/schemas/DecimalMoney'}
        taxFloatDelta: {$ref: '#/components/schemas/SignedDecimalMoney'}
        workingCapitalNeed: {$ref: '#/components/schemas/DecimalMoney'}
        evidence: {type: array, minItems: 1, items: {$ref: '#/components/schemas/DigitalTwinEvidence'}}
        officialSources: {type: array, items: {$ref: '#/components/schemas/OfficialSourceReference'}, description: Resolved only from inherited tax simulation citations}
...
    RegulatoryAuthority:
      type: object
      required: [authorityId, name, officialUrl]
      properties:
        authorityId: {type: string}
        name: {type: string}
        officialUrl: {type: string, format: uri}
    RegulatoryAnswer:
      type: object
      required: [disposition, claims, citations, cutoff_at, policy_version, model_version]
      properties:
        disposition: {type: string, enum: [answered, refused, conflicted, failed]}
        reason: {type: [string, 'null']}
        claims: {type: array, items: {type: object}}
        citations: {type: array, items: {type: object}}
        cutoff_at: {type: string, format: date-time}
        policy_version: {type: string}
        model_version: {type: string}
    RegulatoryChange:
      type: object
      required: [change_request_id, tenant_id, version, status, source_document_version_ids, affected_rule_ids, diff_sha256, created_by, golden_tests_passed]
      properties:
        change_request_id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        version: {type: integer, minimum: 1}
        status: {type: string, enum: [DRAFT, SUBMITTED, APPROVED, REJECTED]}
        source_document_version_ids: {type: array, minItems: 1, items: {type: string}}
        affected_rule_ids: {type: array, items: {type: string}}
        diff_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        created_by: {type: string}
        golden_tests_passed: {type: boolean}
    ShadowTaxDivergence:
      type: object
      required: [divergence_id, reconciliation_id, version, tenant_id, company_tax_id, status, absolute_difference, policy_version, fingerprint, logical_cutoff_at, source_event_ids, evidence]
      properties:
        divergence_id: {type: string, format: uuid}
        reconciliation_id: {type: string, format: uuid}
        version: {type: integer, minimum: 1}
        tenant_id: {type: string, format: uuid}
        company_tax_id: {type: string, pattern: '^\d{14}$'}
        status: {type: string, enum: [MATCHED, MATCHED_WITH_TOLERANCE, DIVERGENT, PENDING_HUMAN_REVIEW, RESOLVED, INVALIDATED]}
        divergence_type: {type: [string, 'null']}
        severity: {type: [string, 'null'], enum: [REVIEW, HIGH, CRITICAL, null]}
        absolute_difference: {$ref: '#/components/schemas/DecimalMoney'}
        policy_version: {type: string}
        fingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}
        logical_cutoff_at: {type: string, format: date-time}
        source_event_ids: {type: array, items: {type: string}}
        evidence:
          type: object
          required: [simulation_ids, rule_ids, memory_ids, official_source_ids, official_source_urls]
          properties:
            simulation_ids: {type: array, items: {type: string}}
            rule_ids: {type: array, items: {type: string}}
            memory_ids: {type: array, items: {type: string}}
            official_source_ids: {type: array, items: {type: string}}
            official_source_urls: {type: array, items: {type: string, format: uri}}
    DecimalMoney:
      type: string
      pattern: '^\d+\.\d{2}$'
      examples: ['100000.00']
    SignedDecimalMoney:
      type: string
      pattern: '^-?\d+\.\d{2}$'
      examples: ['-1250.50']
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string